+++ quant/data_sources/providers/__init__.py
@@ -1,6 +1,7 @@
 from .akshare import AkShareProvider
+from .company_profile_legacy import LegacyCompanyProfileProvider
 from .eastmoney import EastMoneyProvider
 from .fundamentals_legacy import LegacyFundamentalsProvider
 from .tencent import TencentProvider
 
-__all__ = ["AkShareProvider", "EastMoneyProvider", "LegacyFundamentalsProvider", "TencentProvider"]
+__all__ = ["AkShareProvider", "EastMoneyProvider", "LegacyCompanyProfileProvider", "LegacyFundamentalsProvider", "TencentProvider"]